+++ e2e/docker/device-health-oracle/entrypoint.sh
@@ -14,6 +14,10 @@ if [ -z "${DZ_TELEMETRY_PROGRAM_ID}" ]; then
   echo "DZ_TELEMETRY_PROGRAM_ID is not set"
   exit 1
 fi
+if [ -z "${DZ_SIGNER_KEYPAIR}" ]; then
+  echo "DZ_SIGNER_KEYPAIR is not set"
+  exit 1
+fi
 
 # Start Alloy in background if ALLOY_PROMETHEUS_URL is set
 if [ -n "${ALLOY_PROMETHEUS_URL:-}" ]; then
@@ -22,4 +26,4 @@ if [ -n "${ALLOY_PROMETHEUS_URL:-}" ]; then
 fi
 
 # start device-health-oracle
-device-health-oracle -ledger-rpc-url ${DZ_LEDGER_URL} -serviceability-program-id ${DZ_SERVICEABILITY_PROGRAM_ID} -telemetry-program-id ${DZ_TELEMETRY_PROGRAM_ID} -metrics-addr ":2112" -interval ${DZ_INTERVAL:-1m}
+device-health-oracle -ledger-rpc-url ${DZ_LEDGER_URL} -serviceability-program-id ${DZ_SERVICEABILITY_PROGRAM_ID} -telemetry-program-id ${DZ_TELEMETRY_PROGRAM_ID} -signer-keypair ${DZ_SIGNER_KEYPAIR} -metrics-addr ":2112" -interval ${DZ_INTERVAL:-1m}